Split js.hypered.design from hypered.design.
I want to build the site content with Nix. This is the case for
hypered.design.

But I have more work to do for the Next.js, Storybook, and ITCSS parts
of the site. Which is thus now moved to js.hypered.design, and still use
the content of the docs/ directory (exposed through GitHub Pages).