Remove a level of div.
The div was repeated in complete exemples because it was part of the
definition of the `document` function, but also of examples such as
blog-post-1.

It needs to be kept whithin the examples so they match what is defined
in Storybook, so we remove the one in `document`.

We change it slightly so there is less padding on smaller screen (as was
done in the past already).