Fix the normal navbar.
Now a "normal" .c-navbar, which can be used for both a main navigation
bar or a toolbar, doesn't include extra space at the top.

Instead, that extra space is provided now by .c-navbar--main, which is
indeed used in the case of the main navigation bar.

An example page that includes both a main navigation bar and a toolbar
is added. Similarly to the navigation bar, the toolbar uses bare links
instead of proper ul/li or buttons. This should be changed later.