Fix dropdown.
On smaller viewports, the menu items get placed vertically. That's ok
but two problem appeared: the menu items were too close vertically
and the first menu item was a bit more to the left.

The first problem was because the padding on the a element (within the
main menu-item wrapper) has some padding but is inline instead of
inline-block. This is fixed.

The second problem was because all items except the first one had some
left margins. Now they have all some left margins, including the first
one.