Adapt the arrow on a dropdown.
Its placement respective to the dropdown label changes depending on the
viewport dimensions.

We change the value so it's not too close from the label on small screens,
but it's still not ideal (or it becomes too from the label on desktop).
We should revise how its left position is computed).