Fix the layout section on the overview page.
The placeholders CSS rules were missing and are now added.

We make the .wrapper padding thinner, to make it easier to differentiate
padding that is added by the layouting primitive.

We also add the .inset variant.

Finally, we add some content to the .box layout, to be consistent with
the presentation of other primitives.