Downgrade Nixpkgs.
I'll have to revisit this, because it seems some .sty LaTeX files are
missing when building the whole default.nix.

I've regenerated the itcss/nix and emails/nix directory. One weird thing
is that doing so, the selected version of Node was nodejs_16, instead of
the previous nodejs-16_x. I guess there is some mismatch between the
Nixpkgs used in node2nix.sh, and the one specified in nix/sources.json.